finished FUB
except KEY mode, everything is completed for now

+++ src/SRC_HWEP_Sequencer_V2/lib/adc088s052/adc088s052.h
@@ -90,7 +90,7 @@ esp_err_t adc088s052_init(adc088s052_handle_t *out_handle, const adc088s052_conf
  * @param data 		data pointer to store the read data in 
  * @return esp_err_t 
  */
-esp_err_t adc088s052_get_raw(adc088s052_handle_t handle, adc088s052_channel_t ch, uint16_t *data);
+esp_err_t adc088s052_get_raw(adc088s052_handle_t handle, adc088s052_channel_t ch, uint8_t *data);
 
 /**
  * @brief Frees the memory of the handle
